调整数据库SqlFinanceData相关
