+++ src/Senparc.Core/Models/DataBaseModel/AccountPayLog.cs
@@ -2,7 +2,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace Senparc.Scf.Core.Models
+namespace Senparc.Core.Models
 {
     [Serializable]
     public partial class AccountPayLog : EntityBase<int>
