+++ src/Senparc.Core/Models/DataBaseModel/AdminUserInfo.cs
@@ -3,7 +3,7 @@ using Senparc.Scf.Core.Models;
 using Senparc.Scf.Core.Utility;
 using System;
 
-namespace Senparc.Scf.Core.Models
+namespace Senparc.Core.Models
 {
     [Serializable]
     public partial class AdminUserInfo : EntityBase<int>
