+++ src/Senparc.Scf.Core/Cache/BaseCache/BaseDictionary/BaseDictionaryCache.cs
@@ -36,7 +36,7 @@ namespace Senparc.Scf.Core.Cache
         /// <param name="CACHE_KEY"></param>
         /// <param name="db"></param>
         /// <param name="timeOut">单位：分钟。1440为一天。</param>
-        public BaseDictionaryCache(string CACHE_KEY, ISqlClientFinanceData db, int timeOut)
+        public BaseDictionaryCache(string CACHE_KEY, ISqlBaseFinanceData db, int timeOut)
             : base(CACHE_KEY, db, timeOut)
         {
             base.TimeOut = timeOut;
@@ -134,7 +134,7 @@ namespace Senparc.Scf.Core.Cache
         /// <param name="CACHE_KEY"></param>
         /// <param name="db"></param>
         /// <param name="timeOut">单位：分钟。1440为一天。</param>
-        public BaseDictionaryCache(string CACHE_KEY, ISqlClientFinanceData db, int timeOut)
+        public BaseDictionaryCache(string CACHE_KEY, ISqlBaseFinanceData db, int timeOut)
             : base(CACHE_KEY, db)
         {
             base.TimeOut = timeOut;
