+++ src/Senparc.Scf.Core/Cache/BaseCache/BaseDictionary/BaseStringDictionary/BaseStringDictionaryCache.cs
@@ -20,7 +20,7 @@ namespace Senparc.Scf.Core.Cache
         /// <param name="CACHE_KEY"></param>
         /// <param name="db"></param>
         /// <param name="timeOut">单位：分钟。1440为一天。</param>
-        public BaseStringDictionaryCache(string CACHE_KEY, ISqlClientFinanceData db, int timeOut)
+        public BaseStringDictionaryCache(string CACHE_KEY, ISqlBaseFinanceData db, int timeOut)
             : base(CACHE_KEY, db, timeOut)
         {
             base.TimeOut = timeOut;
@@ -37,7 +37,7 @@ namespace Senparc.Scf.Core.Cache
         /// <param name="CACHE_KEY"></param>
         /// <param name="db"></param>
         /// <param name="timeOut">单位：分钟。1440为一天。</param>
-        public BaseStringDictionaryCache(string CACHE_KEY, ISqlClientFinanceData db, int timeOut)
+        public BaseStringDictionaryCache(string CACHE_KEY, ISqlBaseFinanceData db, int timeOut)
             : base(CACHE_KEY, db, timeOut)
         {
             base.TimeOut = timeOut;
