--- src/Senparc.Scf.Core/Models/SenparcEntities.Partial.cs
@@ -1,7 +1,0 @@
-﻿namespace Senparc.Scf.Core.Models
-{
-    public partial class SenparcEntities
-    {
-
-    }
-}
